pgw#1362 wave 3a: the first VALUE pass — and mutation probing refuted three of five deletions
Paul's ruling (2026-08-17, "just because a test exists doesn't mean it's useful") makes VALUE, not
vacuity, the deletion criterion. This is the first wave to run under it, and its headline is a
negative result that changes how every later wave must work.

THE VALUE PASS. All 36 short tests in the procsplit/child-fault cluster were read in full with an
explicit deletion bias, against Paul's classes: N variants of one property, over-specified
internals, trivia, structurally-impossible incidents, deleted systems. Five were condemned.

THE PROBE REFUTED THREE OF THEM. Every condemned test was then checked the only honest way — break
its subject in `src/` and see whether ANYTHING else goes red:

  container_limits() -> {}                    43 passed   <- ONLY guard. RESTORED.
  compile_crash_rows() drops its kind filter  43 passed   <- ONLY guard. RESTORED.
  memory_oom_group never recorded             43 passed   <- ONLY guard. RESTORED.
  describe_exit() names every signal SIGMUTANT  5 FAILED   <- subsumed. Deleted.
  report_worker_fatal ignores an empty addr   29 passed   <- only guard, deliberately given up.

Three of five. The intuition "the broad test already covers this" was wrong 60% of the time, on
tests whose own docstrings undersold them. **A journey-replacement strategy that is not validated
per property will silently shred coverage.** That is now the rule on pgw#1362: no test dies until a
mutation of its subject reddens something else, or the property is explicitly named as given up.

The two deletions that stand:
  * `test_describe_exit_decodes_signals` — a decoder unit test carrying dead code
    (`os.WTERMSIG(9) if False else 9`); a real signal death through `test_signal_death_is_named`
    reddens five tests when the decoder breaks.
  * `test_report_is_a_noop_without_an_orchestrator_address` — uniquely guarded and given up ON
    PURPOSE: the guarded state is unreachable on a pod, which always has an orchestrator address.

A VACUITY FINDING BANKED, NOT FIXED HERE. `test_signal_death_is_named` asserts `memory.max=`,
`memory.current=`, `cpu.max=` and `host_cpu_count=` reach the death dial — and those assertions
still pass with `container_limits()` returning `{}`, because the dial prints the keys regardless.
It checks the strings, not the facts. That is why the restored `container_limits` test is load-
bearing, and it is the next thing to strengthen in this cluster.

THE FOLD. `test_child_faults.py` replaces test_gw640_postmortem, test_worker_fatal_gw640,
test_native_crash_streak_pgw676, test_child_stderr_postmortem_pgw833 and test_pgw714_compile_crash:
5 modules -> 1, 1,105 -> 963 lines, 48 tests in and 46 out. Three more modules leave the mypy
`ignore_errors` burn-down with none added (242 -> 239); `mypy src tests` green over 917 files; all
24 fast-gate lint scripts pass.

Fold probe: `postmortem.record_native_crash` stops de-duplicating attempts of one request -> 4 RED.

Counts, per the ruling's reporting format:
  deleted-as-vacuous 0 · deleted-as-low-value 2 · folded-into-module 46 · kept-focused 3 (restored)
