+++ pyproject.toml
@@ -496,7 +496,6 @@ module = [
   "tests.test_guard_miss_pgw680",
   "tests.test_guard_source_roots_pgw733",
   "tests.test_gw640_disk_only_residency",
-  "tests.test_gw640_postmortem",
   "tests.test_handback_key_axes_pgw1042",
   "tests.test_handler_proof_pgw1199",
   "tests.test_heartbeat_gw613",
@@ -546,7 +545,6 @@ module = [
   "tests.test_modular_checkpoint_dtype_pgw1071",
   "tests.test_modular_hydration_pgw1036",
   "tests.test_multi_group_async_gate_pgw778",
-  "tests.test_native_crash_streak_pgw676",
   "tests.test_nccl_nvls_pgw773",
   "tests.test_numerics_gate_pgw868",
   "tests.test_objectives_pgw654",
@@ -621,7 +619,6 @@ module = [
   "tests.test_warmup_kind_collision_pgw1067",
   "tests.test_warmup_oneof_invariant_th1771",
   "tests.test_weight_free_premise_pgw1173",
-  "tests.test_worker_fatal_gw640",
   "tests.test_wrapper_subclass_virtuality_pgw1198",
   "tests.test_zero_bubble_meter_pgw1154",
   "tests_v2.catalog",
